# Optimize 2 keyword multi-terms aggregation