@sandeshkr419 2024-06-12
For POC, I ran the query the below query for big5 workload and saw 50% reduction in service time.


I tried to benchmark this against eventdata workload since the query time in above big5 workload was too high and I needed a smaller dataset to establish gains, but sadly it doesn't looks like that the change is improving the results. It may actually end up worsening the performance.

| field_i            | Cardinality of field 1: x | field_j                | Cardinality of field 2:y | x * y  | time taken in old code | time taken with change |
|--------------------|----------------------------|------------------------|----------------------------|--------|-----------|-----------|
| httpversion        | 1                          | verb                   | 5                          | 5      | 3563      | 3628      |
| httpversion        | 1                          | useragent.os_name      | 42                         | 42     | 4365      | 4787      |
| httpversion        | 1                          | geoip.country_name     | 190                        | 190    | 7857      | 9140      |
| httpversion        | 1                          | useragent.os           | 203                        | 203    | 6136      | 6953      |
| httpversion        | 1                          | useragent.name         | 208                        | 208    | 5814      | 6586      |
| verb               | 5                          | httpversion            | 1                          | 5      | 3649      | 3442      |
| verb               | 5                          | useragent.os_name      | 42                         | 210    | 4551      | 5047      |
| verb               | 5                          | geoip.country_name     | 190                        | 950    | 8212      | 9450      |
| verb               | 5                          | useragent.os           | 203                        | 1015   | 6532      | 7260      |
| verb               | 5                          | useragent.name         | 208                        | 1040   | 5957      | 6730      |
| useragent.os_name  | 42                         | httpversion            | 1                          | 42     | 4395      | 4905      |
| useragent.os_name  | 42                         | verb                   | 5                          | 210    | 4645      | 5003      |
| useragent.os_name  | 42                         | geoip.country_name     | 190                        | 7980   | 9413      | 11010     |
| useragent.os_name  | 42                         | useragent.os           | 203                        | 8526   | 7457      | 8613      |
| useragent.os_name  | 42                         | useragent.name         | 208                        | 8736   | 7034      | 8227      |
| geoip.country_name | 190                        | httpversion            | 1                          | 190    | 8201      | 9329      |
| geoip.country_name | 190                        | verb                   | 5                          | 950    | 8341      | 9477      |
| geoip.country_name | 190                        | useragent.os_name      | 42                         | 7980   | 9218      | 10865     |
| geoip.country_name | 190                        | useragent.os           | 203                        | 38570  | 11527     | 13095     |
| geoip.country_name | 190                        | useragent.name         | 208                        | 39520  | 10623     | 12593     |
| useragent.os       | 203                        | httpversion            | 1                          | 203    | 6448      | 7058      |
| useragent.os       | 203                        | verb                   | 5                          | 1015   | 6731      | 7294      |
| useragent.os       | 203                        | useragent.os_name      | 42                         | 8526   | 7391      | 8692      |
| useragent.os       | 203                        | geoip.country_name     | 190                        | 38570  | 11351     | 13005     |
| useragent.os       | 203                        | useragent.name         | 208                        | 42224  | 8876      | 10156     |
| useragent.name     | 208                        | httpversion            | 1                          | 208    | 5975      | 6651      |
| useragent.name     | 208                        | verb                   | 5                          | 1040   | 6155      | 6818      |
| useragent.name     | 208                        | useragent.os_name      | 42                         | 8736   | 6934      | 8149      |
| useragent.name     | 208                        | geoip.country_name     | 190                        | 39520  | 10807     | 12583     |
| useragent.name     | 208                        | useragent.os           | 203                        | 42224  | 8827      | 10220     |